requirementsLocal.txt
pymongo~=4.1.1
Flask~=2.1.2
MarkupSafe==2.1.1
flask_cors
requests==2.28.0
pyjwt[crypto]
boto3
numpy
argon2-cffi
